refactor: rename file
In `fast-check`, the library we are using for property-based tests, data-generators are called arbitraries. In order to match this convention, renamed the module which generates domain-names to `domain-name.arb.ts`.
