+++ test/package-url.test.ts
@@ -1,6 +1,6 @@
 import { isPackageUrl } from "../src/types/package-url";
 import fc from "fast-check";
-import { arbDomainName } from "./domain-name.gen";
+import { arbDomainName } from "./domain-name.arb";
 
 describe("package-url", () => {
   describe("validation", () => {
